+++ R/ggpar.R
@@ -29,6 +29,7 @@ NULL
 #'   font.x = "bold", to change only font face.
 #' @param font.title,font.subtitle alias of font.submain and font.submain,
 #'   respectively.
+#' @param font.family character vector specifying font family.
 #' @param xlim,ylim a numeric vector of length 2, specifying  x and y axis
 #'   limits (minimum and maximum), respectively. e.g.: ylim = c(0, 50).
 #' @param xscale,yscale x and y axis scale, respectively. Allowed values are one
@@ -39,14 +40,16 @@ NULL
 #'   c("top", "bottom", "left", "right", "none"). To remove the legend use
 #'   legend = "none". Legend position can be also specified using a numeric
 #'   vector c(x, y); see details section.
-#' @param legend.title legend title.
+#' @param legend.title legend title, e.g.: \code{legend.title = "Species"}. Can
+#'   be also a list, \code{legend.title = list(color = "Species", linetype = "Species",
+#'   shape = "Species")}.
 #' @param font.legend legend text font style; e.g.: font.legend = c(10, "plain",
 #'   "black").
 #' @param ticks logical value. Default is TRUE. If FALSE, hide axis tick marks.
 #' @param tickslab logical value. Default is TRUE. If FALSE, hide axis tick
 #'   labels.
-#' @param font.tickslab,font.xtickslab,font.ytickslab Font style (size, face, color) for tick labels, e.g.:
-#'   c(14, "bold", "red").
+#' @param font.tickslab,font.xtickslab,font.ytickslab Font style (size, face,
+#'   color) for tick labels, e.g.: c(14, "bold", "red").
 #' @param x.text.angle,y.text.angle Numeric value specifying the rotation angle
 #'   of x and y axis tick labels, respectively. Default value is NULL. For
 #'   vertical x axis texts use x.text.angle = 90.
@@ -145,7 +148,7 @@ ggpar <- function(p, palette = NULL, gradient.cols = NULL,
                   main = NULL, submain = NULL, caption = NULL, xlab = NULL, ylab = NULL,
                   title = NULL, subtitle = NULL,
                   font.main = NULL, font.submain = NULL, font.x = NULL, font.y = NULL, font.caption = NULL,
-                  font.title = NULL, font.subtitle = NULL,
+                  font.title = NULL, font.subtitle = NULL, font.family = "",
                   xlim = NULL, ylim = NULL,
                   xscale = c("none", "log2", "log10", "sqrt"),
                   yscale = c("none", "log2", "log10", "sqrt"),
@@ -195,6 +198,8 @@ ggpar <- function(p, palette = NULL, gradient.cols = NULL,
                      font.main, font.x, font.y,
                    submain = submain, caption = caption, font.submain = font.submain, font.caption = font.caption)
         p <- .set_orientation(p, orientation)
+        if(font.family != "")
+          p <- p + theme(text = element_text(family = font.family))
         list.plots[[i]] <- p
     }