+++ R/ggparagraph.R
@@ -75,6 +75,12 @@ splitString <- function(text) {
   newstring
 }
 
+#' @method drawDetails splitText
+#' @rdname ggparagraph
+#' @param x a grid grob
+#' @param recording a logical value indicating whether a grob is being added to
+#'   the display list or redrawn from the display list.
+#' @export
 drawDetails.splitText <- function(x, recording) {
   grid::grid.text(splitString(x$text),
             x=0, y=1, just=c("left", "top"))